requirements.txt
discord.py
requests
pyyaml
flake8
black
bandit